Additional features and functionality:
based on issue inasafe/inasafe-realtime#35

Celery:
- Limit prefetch multiplier explicitly to 1 (avoid hanging worker)

Earthquake:
- Add MMI Output FileField
- Add download button for MMI Output zip
- Auto zoom to first earthquake in the table

Flood:
- Fix signals to recalculate total affected population and flooded RW
- Add Data Source Field, since it needs to support PetaBencana
- Add code branching to handle old hazard data and new PetaBencana
- Add download button for Hazard File and Impact Layers
- Change download button icon to follow Earthquake pattern
- Auto zoom to first flood in the table

Ash:
- Add Impact Files FileField
- Add download button for Impact Files zip
- Change download button icon to follow Earthquake pattern
- Auto zoom to first ash event in the table

+++ django_project/realtime/views/flood.py
@@ -140,8 +140,10 @@ class FloodDetail(mixins.RetrieveModelMixin, mixins.UpdateModelMixin,
             event_id = request.data['event_id']
             if event_id:
                 event = Flood.objects.get(event_id=event_id)
-                event.hazard_layer.delete()
-                event.impact_layer.delete()
+                if 'hazard_layer' in request.FILES and event.hazard_layer:
+                    event.hazard_layer.delete()
+                if 'impact_layer' in request.FILES and event.impact_layer:
+                    event.impact_layer.delete()
                 retval = self.update(request, partial=True, *args, **kwargs)
                 return retval
             else:
